+++ test/integration/concerns/metadata_templates_test.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+require 'test_helper'
+
+class MetadataTemplatesTest < ActionDispatch::IntegrationTest
+  test 'metadata_templates_path raises NotImplementedError when not overridden' do
+    controller_class = Class.new(ApplicationController) do
+      include MetadataTemplateActions
+    end
+
+    assert_raises(NotImplementedError) do
+      controller_class.new.send(:metadata_templates_path)
+    end
+  end
+end
